Employee entry register application
This is demo application for keeping employee entry details. This demo application contains and maintain arrival employee details ( name and phonenumber ). This registration module contains we can add the emp details and delete the emp details base on the requirements. This application using COREDATA to maintain the record and its developed by SWIFT 5.2.3